rename config utils
